+++ lib/src/debug/trace.dart
@@ -15,10 +15,10 @@ import '../shared/types.dart';
 ///
 /// produces the following output:
 ///
-///     Instance of 'SequenceParser'
+///     Instance of 'SequenceParser<dynamic>'
 ///       Instance of 'CharacterParser'[letter expected]
 ///       Success[1:2]: f
-///       Instance of 'PossessiveRepeatingParser'[0..*]
+///       Instance of 'PossessiveRepeatingParser<String>'[0..*]
 ///         Instance of 'CharacterParser'[letter or digit expected]
 ///         Success[1:3]: 1
 ///         Instance of 'CharacterParser'[letter or digit expected]